tools/pyproject.toml
[tool.poetry]
name = "pdf-image-analyzer"
version = "0.1.0"
description = "Tools to convert PDF pages to images and analyze them using LLM."
authors = ["Your Name <you@example.com>"] # 作者情報を適宜変更してください
readme = "README.md"
license = "MIT" # または適切なライセンスを選択
package-mode = false

[tool.poetry.dependencies]
python = "^3.10" # 要求するPythonのバージョン
pdf2image = "^1.17.0"
Pillow = "^10.3.0" # pdf2image や PIL.Image のために明示
pandas = "^2.2.3"
beautifulsoup4 = "^4.13.4"
tqdm = "^4.67.1"
requests = "^2.32.4"
tenacity = "^9.1.2"
langchain = "^0.3.0"
langchain-google-genai = "^2.0.0"
langchain-anthropic = "^0.3.0"
langchain-openai = "^0.3.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
# 開発時に便利なツールがあればここに追加 (例: black, flake8, pytest)
# black = "^24.4.2"
# flake8 = "^7.0.0"
python-dotenv = "^1.0.1" # .env ファイルを使う場合に便利
pandas-stubs = "^2.2.3.250308"
pytest = "^8.3.5"
ruff = "^0.11.9"
pyright = "^1.1.400"
lefthook = "^1.11.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# スクリプトをコマンドとして実行可能にする場合 (例)
# pdf-to-images = "pdf_image_analyzer.pdf_to_images:main_cli_function"
# analyze-images = "pdf_image_analyzer.analyze_image:main_cli_function"
# 上記を使用するには、現在のスクリプトを関数ベースにリファクタリングする必要があります

[tool.ruff]
line-length = 120
extend-include = ["*.ipynb"]
exclude = []

[tool.ruff.lint]
# チェックしたいルール
# https://docs.astral.sh/ruff/rules/
select = [
    "F",  # pyflakes
    "E",  # pycodestyle
    "W",  # pycodestyle warnings
    "B",  # flake8-bugbear
    "SIM", # flake8-simplify
    "I",  # isort
    "UP", # pyupgrade
    "N",  # pep8-naming
]
# 無視したいルール
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", # asserts allowed in tests...
    "SLF",
    "TRY",
    "ANN", # Missing type
    "D", # Documentation
    "PLR2004", # Magic value used in comparison, ...
]

[tool.pyright]
# VSCodeのデフォルトのモードにあわせる
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
